refactor: separate codebase
